+++ src/components/stepforms/index.ts
@@ -3,6 +3,7 @@ import './AppendStepForm.vue';
 import './ArgminStepForm.vue';
 import './ArgmaxStepForm.vue';
 import './ConcatenateStepForm.vue';
+import './ConvertStepForm.vue';
 import './DeleteColumnStepForm.vue';
 import './DuplicateColumnStepForm.vue';
 import './DomainStepForm.vue';